requirements.txt
# Server
fastapi
uvicorn
scikit-learn
pandas
joblib
python-dotenv
python-multipart
psycopg2-binary
bcrypt
email-validator
PyJWT
pydantic[email]

# Client
streamlit
requests